Refactoring to reduce class complexity and clarify filenames
