+++ Dns/ZoneProvider/IPProbe/Host.cs
@@ -2,13 +2,10 @@
 {
     using System.Collections.Generic;
 
-    public partial class IPProbeZoneProvider
+    internal class Host
     {
-        internal class Host
-        {
-            internal string Name { get; set; }
-            internal AvailabilityMode AvailabilityMode { get; set; }
-            internal List<Target> AddressProbes = new List<Target>();
-        }
+        internal string Name { get; set; }
+        internal AvailabilityMode AvailabilityMode { get; set; }
+        internal List<Target> AddressProbes = new List<Target>();
     }
 }
